app: split preferences code into it own file; PF adjustments
LLM-assisted commit
